Segregated checker tests
