+++ os/test/src/TestUtil.scala
@@ -1,6 +1,7 @@
 package test.os
 
 import utest.framework.TestPath
+
 import java.io.IOException
 import java.nio.file._
 import java.nio.file.attribute.BasicFileAttributes
